refactor(portal): split admin billing route
